+++ data/import_scripts/test_local_import.txt
@@ -1,0 +1,9 @@
+
+### check for unique ain ###
+select ain, count(*)
+  from rawparceltable
+ group by ain
+having count(*) > 1;
+
+
+###
